Refactoring: scss파일 정리 및 컴포넌트 분리 - className 재정의 및 Sass nesting 활용 - Feed.js, Aside.js  반복되는 코드 .map()으로 수정 - Header.js, Aside.js 컴포넌트 분리
